requirements.txt
httpx
pyyaml
python-dotenv
openai
psycopg2-binary
pydantic